+++ Pod/Classes/SEGUrbanAirshipIntegration.m
@@ -1,27 +1,4 @@
-/*
- Copyright 2009-2016 Urban Airship Inc. All rights reserved.
-
- Redistribution and use in source and binary forms, with or without
- modification, are permitted provided that the following conditions are met:
-
- 1. Redistributions of source code must retain the above copyright notice, this
- list of conditions and the following disclaimer.
-
- 2. Redistributions in binary form must reproduce the above copyright notice,
- this list of conditions and the following disclaimer in the documentation
- and/or other materials provided with the distribution.
-
- THIS SOFTWARE IS PROVIDED BY THE URBAN AIRSHIP INC ``AS IS'' AND ANY EXPRESS OR
- IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF
- MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO
- EVENT SHALL URBAN AIRSHIP INC OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
- INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
- BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
- LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE
- OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF
- ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
+/* Copyright Airship and Contributors */
 
 #import "SEGUrbanAirshipIntegration.h"
 #import "SEGUrbanAirshipAutopilot.h"
@@ -32,7 +9,7 @@
 #define kUrbanAirshipAppSecret @"appSecret"
 
 /**
- * Urban Airship Segment integration.
+ * Airship Segment integration.
  */
 @implementation SEGUrbanAirshipIntegration
 
@@ -47,7 +24,31 @@
 
 - (void)identify:(SEGIdentifyPayload *)payload {
     [UAirship namedUser].identifier = payload.userId;
-    [[UAirship push] updateRegistration];
+
+    NSMutableArray *addTags = [NSMutableArray array];
+    NSMutableArray *removeTags = [NSMutableArray array];
+
+    for (NSString *key in payload.traits) {
+        id value = payload.traits[key];
+        if (![value isKindOfClass:[NSNumber class]]) {
+            continue;
+        }
+
+        CFTypeID type = CFGetTypeID((__bridge CFTypeRef)(value));
+        if (type != CFBooleanGetTypeID()) {
+            continue;
+        }
+
+        if ([value boolValue]) {
+            [addTags addObject:key];
+        } else {
+            [removeTags addObject:key];
+        }
+    }
+
+    [[UAirship namedUser] addTags:addTags group:@"segment-integration"];
+    [[UAirship namedUser] removeTags:removeTags group:@"segment-integration"];
+    [[UAirship namedUser] updateTags];
 }
 
 - (void)screen:(SEGScreenPayload *)payload {
@@ -56,6 +57,7 @@
 
 - (void)track:(SEGTrackPayload *)payload {
     UACustomEvent *customEvent = [UACustomEvent eventWithName:payload.event];
+    customEvent.transactionID = @"cdp";
 
     NSNumber *value = [SEGUrbanAirshipIntegration extractRevenue:payload.properties withKey:@"revenue"];
     if (!value) {